// src/app/shared/models/ee-conversion-cost.model.ts
export class EeConversionCost {
  conversionCostId: number | null;
  conversionCostType: string;
  projectInfoId: number | null;
  partInfoId: number | null;
  panelLength: number | null;
  panelWidth: number | null;
  pcbLength: number | null;
  pcbWidth: number | null;
  pcbThickness: number | null;
  panelKeepout: number | null;
  rowValue: number | null;
  columnValue: number | null;
  noOfPcbPanel: number | null;
  dePanelization: boolean | null;
  dePanelizationSelection: number | null;
  mouseBiteSelected: boolean | null;
  mouseBite: number | null;
  vGroveSelected: boolean | null;
  vgrove: number | null;
  smdSelection: number | null;
  selectedMachineTopSide: string;
  selectedMachineBottomSide: string;
  top_Complexity: number | null;
  top_LeadParts: number | null;
  top_ICParts: number | null;
  top_ComplexParts: number | null;
  top_TotalSMDParts: number | null;
  top_ICPlacement: number | null;
  top_LeadPlacement: number | null;
  top_ComplexPlacement: number | null;
  top_TotalSMDPlacements: number | null;
  bottom_Complexity: number | null;
  bottom_LeadParts: number | null;
  bottom_ICParts: number | null;
  bottom_ComplexParts: number | null;
  bottom_TotalSMDParts: number | null;
  bottom_ICPlacement: number | null;
  bottom_LeadPlacement: number | null;
  bottom_ComplexPlacement: number | null;
  bottom_TotalSMDPlacements: number | null;
  waveSolderpartTypes: number | null;
  tHPlacements: number | null;
  manualSolderPartTypes: number | null;
  manualSolderPlacements: number | null;
  selectiveSolderPartType: number | null;
  noOfPems: number | null;
  selectiveSolderJoints: number | null;
  totalTHSolderJoints: number | null;
  holesToBeMasked: number | null;
  manualSolderJoints: number | null;
  nodeCount: number | null;
  highLevelAssemblyParts: number | null;
  pressFitConnectorCount: number | null;
  gluingOperation: number | null;
  modification: number | null;
  inspectionThroughHoleOpert: number | null;
  icProgramming: number | null;
  axi: number | null;
  manualConformalCoating: number | null;
  mechanicIntegration: number | null;
  manualXRayInspection: number | null;
  myProperty: number | null;
  finalInspection: number | null;
  finalTesting: number | null;
  machineConformalCoating: number | null;
  packing: number | null;
  flyingProbeTest: number | null;
  automaticOpticalInspection: number | null;
  mechHoles: number | null;
  hotBarSoldering: number | null;
}

export class EeConversionCostResult {
  conversionCostResultId: number | null;
  conversionCostId: number | null;
  projectInfoId: number | null;
  partInfoId: number | null;

  process: string | null;
  time: number | null;
  mhr: number | null;
  lhr: number | null;
  noOfOperators: number | null;
  conversionCost: number | null;
}